requirements.txt
albumentations==1.0.3
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cosine-annealing-warmup==2.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
easydict==1.9
efficientnet-pytorch==0.6.3
entrypoints==0.3
grad-cam==1.3.1
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.6.1
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaggle==1.5.12
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
munch==2.5.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.0
numpy==1.19.5
opencv-python==4.5.4.60
opencv-python-headless==4.5.3.56
packaging==21.0
paddlepaddle-gpu==2.2.0
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.0
pkg-resources==0.0.0
pretrainedmodels==0.7.4
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.25.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
segmentation-models-pytorch==0.2.0
Send2Trash==1.7.1
Shapely==1.7.1
six==1.16.0
tabulate==0.8.9
tensorboardX==2.4
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2020.9.3
timm==0.4.12
torch==1.7.1
torchvision==0.8.2
tornado==6.1
tqdm==4.61.1
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.5.0
